+++ modules/logic/cognitive_state_analyzer.py
